Creation of Main.py the beginning of a Journey
A Python file that prints text, 

AKA It Echo's text if you like Php. Or if you like JavaScript it logs text in the console. Or if you like HTML it creates a paragraph with text. Or if you like some other language figure out the equivalent of print in the language you desire.